+++ mle_toolbox/hyperopt/hyperopt_base.py
@@ -10,8 +10,7 @@ from pprint import pformat
 from .hyper_logger import HyperoptLogger
 from ..job import JobQueue
 from ..utils import load_json_config, print_framed
-from mle_logging.merge import merge_hdf5_files
-from mle_logging.load import load_meta_log
+from mle_logging import merge_config_logs, load_meta_log
 
 
 class BaseHyperOptimisation(object):
@@ -257,9 +256,15 @@ class BaseHyperOptimisation(object):
         if not self.hyper_log.no_results_logging:
             while True:
                 try:
-                    meta_eval_log = self.get_meta_eval_log(
+                    merge_config_logs(
+                        self.experiment_dir,
                         self.hyper_log.all_run_ids + run_ids
                     )
+                    # Load in meta-results log with values meaned over seeds
+                    meta_log_fname = os.path.join(self.experiment_dir,
+                                                  "meta_log.hdf5")
+                    meta_eval_log = load_meta_log(meta_log_fname,
+                                                  aggregate_seeds=True)
                     break
                 except Exception:
                     time.sleep(1)
@@ -334,35 +339,3 @@ class BaseHyperOptimisation(object):
             all_run_ids.append(run_id)
 
         return config_fnames_batch, all_run_ids
-
-    def get_meta_eval_log(self, all_run_ids):
-        """Scavenge the experiment dictonaries & load in logs."""
-        all_folders = [x[0] for x in os.walk(self.experiment_dir)][1:]
-        # Get rid of timestring in beginning & collect all folders/hdf5 files
-        hyperp_results_folder = []
-        # Need to make sure that run_ids & experiment folder match!
-        for run_id in all_run_ids:
-            for f in all_folders:
-                if f[len(self.experiment_dir) + 9 :] == run_id:
-                    hyperp_results_folder.append(f)
-                    continue
-
-        # Collect all paths to the .hdf5 file
-        log_paths = []
-        for i in range(len(hyperp_results_folder)):
-            log_d_t = os.path.join(hyperp_results_folder[i], "logs/")
-            for file in os.listdir(log_d_t):
-                fname, fext = os.path.splitext(file)
-                if file.endswith(".hdf5") and fname in all_run_ids:
-                    log_paths.append(os.path.join(log_d_t, file))
-
-        # Merge individual run results into a single hdf5 file
-        meta_log_fname = os.path.join(self.experiment_dir, "meta_log.hdf5")
-
-        assert len(log_paths) == len(all_run_ids)
-
-        merge_hdf5_files(meta_log_fname, log_paths, file_ids=all_run_ids)
-
-        # Load in meta-results log with values meaned over seeds
-        meta_eval_logs = load_meta_log(meta_log_fname, aggregate_seeds=True)
-        return meta_eval_logs