+++ mle_toolbox/pbt/pbt_manager.py
@@ -11,7 +11,7 @@ from tqdm import tqdm
 
 from ..utils import load_json_config
 from mle_toolbox.job import Job
-from mle_logging import load_run_log
+from mle_logging import load_log
 
 from .pbt_logger import PBT_Logger
 from .explore import ExplorationStrategy