+++ mle_toolbox/utils/load_hyper_log.py
@@ -1,4 +1,4 @@
-import pickle
+import pickle5 as pickle
 import pandas as pd
 import numpy as np
 from typing import Union, List